refactor: add LanguageSwitcher component for language selection
